Use relative submodule URLs
I prefer a relative link here so that my fork of this project will use
my fork of its submodules too.  Otherwise, there's really no way for me
to make changes in a submodule and have them picked up in this project.

The down side of relative URLs, though, is that they effectively
*require* making forks of submodules to go along with forks of the
containing project.  I don't know if there's a simple way to resolve
that without giving up the benefits of relative URLs.  For right now, at
least, I do need forks of submodules too, so relative URLs are the right
choice.