Avoid a warning about using EqPred uninitialized
With enough warning flags on, some compilers will warn that EqPred may
be uninitialized in the call to createNarrowingFn at the very the end
of SymbolicRangeAnalysis::handleBranch.  However, this cannot actually
happen:

In order for EqPred to be used uninitialized in the createNarrowingFn
call, we would need to have an integer comparison instruction ICI for
which ICmpInst::isEquality, CmpInst::isTrueWhenEqual, and
CmpInst::isFalseWhenEqual all return false.  Inspecting
CmpInst::isTrueWhenEqual and CmpInst::isFalseWhenEqual closely, I find
that for both of these to return false, the comparison predicate would
need to be FCMP_OEQ, FCMP_OGE, FCMP_OLE, FCMP_ORD, FCMP_UNO, FCMP_UGT,
FCMP_ULT, or FCMP_UNE.  However, notice that these are all
floating-point comparison predicates.  Those should only appear as
predicates for FCmpInst instructions, never for the ICmpInst
instructions we'll have at this point in the code.

Thus, for integer comparison instructions, either
CmpInst::isTrueWhenEqual or CmpInst::isFalseWhenEqual must be true,
which means we must always initialize EqPred.  The warning is
spurrious.  This commit changes the last "if" into an assert to
document what we know must be true.